Add files via upload
This is a remade version of the original start menu. I added particle systems to look like stars in the start screen. I could not figure out how to program the buttons to do things whenever you click them. I also added a zilch script to restart the current level whenever R is pressed, however you will have to add the component in the level settings.