Added clamp to player velocity, movement particle effects, and fixed up a few things.
previusly the player could accelerate forever,I added a max and min speed for each axis. The player nnow emits a particle effect when moving. The player's laser shots now have the player's velocity added to them, so they always appear to go the same speed. Tweaked some of the numbers on the laser shots, enemies, and hives.